requirements.txt
streamlit
mecab-python3
